+++ packages/icons/src/components/AppleIcon.tsx
@@ -5,8 +5,10 @@ export interface AppleIconProps {
   isDisabled?: boolean;
   variant?: 'outline' | 'solid';
 }
+
 const AppleIcon: FC<AppleIconProps> = (props) => {
-  const { isDisabled, variant } = props;
+  const { isDisabled, variant = 'outline' } = props;
+
   return (
     <Icon
       width="24px"
